+++ instrumentation/logback/logback-appender-1.0/library/src/test/java/io/opentelemetry/instrumentation/logback/appender/v1_0/internal/LoggingEventMapperTest.java
@@ -103,7 +103,7 @@ class LoggingEventMapperTest {
     LoggingEventMapper mapper =
         LoggingEventMapper.builder()
             .setMdcAttributes(
-                MdcAttributeSelectors.create(
+                AttributeSelectors.create(
                     IncludeExclude.builder().setExcluded(singletonList("*secret*")).build()))
             .build();
     Map<String, String> contextData = new HashMap<>();
@@ -122,7 +122,7 @@ class LoggingEventMapperTest {
     LoggingEventMapper mapper =
         LoggingEventMapper.builder()
             .setMdcAttributes(
-                MdcAttributeSelectors.create(
+                AttributeSelectors.create(
                     IncludeExclude.builder()
                         .setIncluded(singletonList("request-*"))
                         .setExcluded(singletonList("*-secret"))
@@ -143,7 +143,7 @@ class LoggingEventMapperTest {
   void testEmptySelectorCapturesNothing() {
     LoggingEventMapper mapper =
         LoggingEventMapper.builder()
-            .setMdcAttributes(MdcAttributeSelectors.create(IncludeExclude.builder().build()))
+            .setMdcAttributes(AttributeSelectors.create(IncludeExclude.builder().build()))
             .build();
     Map<String, String> contextData = new HashMap<>();
     contextData.put("key1", "value1");
@@ -257,7 +257,7 @@ class LoggingEventMapperTest {
   }
 
   private static Predicate<String> include(String... patterns) {
-    return MdcAttributeSelectors.create(
+    return AttributeSelectors.create(
         IncludeExclude.builder().setIncluded(asList(patterns)).build());
   }
 }